Add MbusStreamParser that handles continous stream of data
The parser has internal state that compares incoming data to the expected
frame start byte 0x7E, and pushData() return true if and only if the end
of a frame is reached. pushData() will overwrite the internal buffer, so
the result from getFrame() will be invalidated when pushData() is called
again.